add logic to unlink associated car when deleting parking location

+++ CPT331.Carbuds.Api/Services/ParkingLocationService.cs
@@ -9,57 +9,57 @@ using Microsoft.Extensions.Configuration;
 
 namespace CPT331.Carbuds.Api.Services
 {
-    public interface IParkingLocationService
+  public interface IParkingLocationService
+  {
+    Task<List<ParkingLocation>> ListAllParkingLocations();
+    Task<ParkingLocation> GetParkingLocation(string parkingLocationId);
+    Task<List<ParkingLocation>> ListAvailableParkingLocations();
+    Task<bool> AddUpdateParkingLocation(ParkingLocation record);
+    Task<bool> DeleteParkingLocation(string parkingLocationUuid);
+  }
+
+  public class ParkingLocationService : IParkingLocationService
+  {
+    private IAmazonDynamoDB _dynamoDb;
+    private IConfiguration _config;
+    private IUtilityService _utils;
+    private IParkingAllocationService _plService;
+
+    // constructor
+    public ParkingLocationService(IAmazonDynamoDB dynamoDb, IConfiguration config, IUtilityService utils, IParkingAllocationService plService)
     {
-        Task<List<ParkingLocation>> ListAllParkingLocations();
-        Task<ParkingLocation> GetParkingLocation(string parkingLocationId);
-        Task<List<ParkingLocation>> ListAvailableParkingLocations();
-        Task<bool> AddUpdateParkingLocation(ParkingLocation record);
-        Task<bool> DeleteParkingLocation(string parkingLocationUuid);
+      _dynamoDb = dynamoDb;
+      _config = config;
+      _utils = utils;
+      _plService = plService;
     }
 
-    public class ParkingLocationService : IParkingLocationService
+    // method to retrieve a list of parking locations from database
+    public async Task<List<ParkingLocation>> ListAllParkingLocations()
     {
-        private IAmazonDynamoDB _dynamoDb;
-        private IConfiguration _config;
-        private IUtilityService _utils;
-        private IParkingAllocationService _plService;
+      var parkingLocationList = new List<ParkingLocation>();
+      ScanRequest scanReq = new ScanRequest()
+      {
+        TableName = _config.GetValue<string>("DynamoDb:Tablenames:ParkingLocations")
+      };
 
-        // constructor
-        public ParkingLocationService(IAmazonDynamoDB dynamoDb, IConfiguration config, IUtilityService utils, IParkingAllocationService plService)
-        {
-            _dynamoDb = dynamoDb;
-            _config = config;
-            _utils = utils;
-            _plService = plService;
-        }
+      var dbResult = await _dynamoDb.ScanAsync(scanReq);
+      foreach (var item in dbResult.Items)
+      {
+        parkingLocationList.Add(_utils.ToObjectFromDynamoResult<ParkingLocation>(item));
+      }
 
-        // method to retrieve a list of parking locations from database
-        public async Task<List<ParkingLocation>> ListAllParkingLocations()
-        {
-            var parkingLocationList = new List<ParkingLocation>();
-            ScanRequest scanReq = new ScanRequest()
-            {
-                TableName = _config.GetValue<string>("DynamoDb:Tablenames:ParkingLocations")
-            };
+      return parkingLocationList;
+    }
 
-            var dbResult = await _dynamoDb.ScanAsync(scanReq);
-            foreach (var item in dbResult.Items)
-            {
-                parkingLocationList.Add(_utils.ToObjectFromDynamoResult<ParkingLocation>(item));
-            }
-
-            return parkingLocationList;
-        }
-
-        public async Task<ParkingLocation> GetParkingLocation(string parkingLocationId)
-        {
-            QueryRequest query = new QueryRequest()
-            {
-                TableName = _config.GetValue<string>("DynamoDb:Tablenames:ParkingLocations"),
-                ReturnConsumedCapacity = "TOTAL",
-                KeyConditionExpression = "#Uuid = :v_Uuid",
-                ExpressionAttributeValues = new Dictionary<string, AttributeValue>()
+    public async Task<ParkingLocation> GetParkingLocation(string parkingLocationId)
+    {
+      QueryRequest query = new QueryRequest()
+      {
+        TableName = _config.GetValue<string>("DynamoDb:Tablenames:ParkingLocations"),
+        ReturnConsumedCapacity = "TOTAL",
+        KeyConditionExpression = "#Uuid = :v_Uuid",
+        ExpressionAttributeValues = new Dictionary<string, AttributeValue>()
                 {
                     {
                         ":v_Uuid",
@@ -69,49 +69,64 @@ namespace CPT331.Carbuds.Api.Services
                         }
                     }
                 },
-                ExpressionAttributeNames = new Dictionary<string, string>()
+        ExpressionAttributeNames = new Dictionary<string, string>()
                 {
                   {
                     "#Uuid" , "Uuid"
                   }
                 }
-            };
+      };
 
-            var queryResult = await _dynamoDb.QueryAsync(query);
-            return queryResult.Items.Any() ? _utils.ToObjectFromDynamoResult<ParkingLocation>(queryResult.Items.First()) : null;
-        }
+      var queryResult = await _dynamoDb.QueryAsync(query);
+      return queryResult.Items.Any() ? _utils.ToObjectFromDynamoResult<ParkingLocation>(queryResult.Items.First()) : null;
+    }
 
-        public async Task<List<ParkingLocation>> ListAvailableParkingLocations()
-        {
-            var allParkingLocations = await ListAllParkingLocations();
-            var allAllocations = await _plService.ListAllParkingAllocations();
+    public async Task<List<ParkingLocation>> ListAvailableParkingLocations()
+    {
+      var allParkingLocations = await ListAllParkingLocations();
+      var allAllocations = await _plService.ListAllParkingAllocations();
 
-            return allParkingLocations.Except(allParkingLocations.Join(allAllocations, p => p.Uuid, a => a.LocationUuid, (p, a) => p)).ToList();
-        }
+      return allParkingLocations.Except(allParkingLocations.Join(allAllocations, p => p.Uuid, a => a.LocationUuid, (p, a) => p)).ToList();
+    }
 
-        public async Task<bool> AddUpdateParkingLocation(ParkingLocation record)
-        {
-            var putReq = new PutItemRequest()
-            {
-                TableName = _config.GetValue<string>("DynamoDb:Tablenames:ParkingLocations"),
-                Item = _utils.ToDynamoAttributeValueDictionary<ParkingLocation>(record)
-            };
-            var response = await _dynamoDb.PutItemAsync(putReq);
-            return true;
-        }
+    public async Task<bool> AddUpdateParkingLocation(ParkingLocation record)
+    {
+      var putReq = new PutItemRequest()
+      {
+        TableName = _config.GetValue<string>("DynamoDb:Tablenames:ParkingLocations"),
+        Item = _utils.ToDynamoAttributeValueDictionary<ParkingLocation>(record)
+      };
+      var response = await _dynamoDb.PutItemAsync(putReq);
+      return true;
+    }
 
-        public async Task<bool> DeleteParkingLocation(string parkingLocationUuid)
-        {
-            var delReq = new DeleteItemRequest()
-            {
-                TableName = _config.GetValue<string>("DynamoDb:Tablenames:ParkingLocations"),
-                Key = new Dictionary<string, AttributeValue>() {
+    public async Task<bool> DeleteParkingLocation(string parkingLocationUuid)
+    {
+      var delReq = new DeleteItemRequest()
+      {
+        TableName = _config.GetValue<string>("DynamoDb:Tablenames:ParkingLocations"),
+        Key = new Dictionary<string, AttributeValue>() {
             { "Uuid", new AttributeValue { S = parkingLocationUuid } },
         }
-            };
+      };
 
-            var response = await _dynamoDb.DeleteItemAsync(delReq);
-            return true;
-        }
+      var existingParkingAllocation = (await _plService.ListAllParkingAllocations()).Where(p => p.LocationUuid == parkingLocationUuid).FirstOrDefault();
+      if(existingParkingAllocation != null)
+      {
+        var delPlAllocationReq = new DeleteItemRequest()
+        {
+          TableName = _config.GetValue<string>("DynamoDb:Tablenames:CarParkingAllocations"),
+          Key = new Dictionary<string, AttributeValue>() {
+            { "CarUuid", new AttributeValue { S = existingParkingAllocation.CarUuid } },
+            { "LocationUuid", new AttributeValue { S = existingParkingAllocation.LocationUuid } },
+          }
+        };
+        var plAllocationDeleted = await _dynamoDb.DeleteItemAsync(delPlAllocationReq);
+      }
+      
+
+      var response = await _dynamoDb.DeleteItemAsync(delReq);
+      return true;
     }
+  }
 }
